+++ examples/nuxtjs-client-example/components/Tutorial.vue
@@ -1,15 +1,7 @@
 <!-- Please remove this file from your project -->
 <template>
   <div
-    class="
-      relative
-      flex
-      items-top
-      justify-center
-      min-h-screen
-      bg-gray-100
-      sm:items-center sm:pt-0
-    "
+    class="relative flex items-top justify-center min-h-screen bg-gray-100 sm:items-center sm:pt-0"
   >
     <link
       href="https://cdn.jsdelivr.net/npm/tailwindcss@2.1.2/dist/tailwind.min.css"