+++ src/plugin.ts
@@ -1,6 +1,11 @@
-import { PiniaPlugin, PiniaPluginContext } from 'pinia'
+import {
+  PiniaPlugin,
+  PiniaPluginContext,
+  StateTree,
+  SubscriptionCallbackMutation,
+} from 'pinia'
 import * as shvl from 'shvl'
-import { CommonOptions } from '.'
+import { CommonOptions, IStorage } from '.'
 import { PluginOptions } from './type'
 
 function getOption<K extends keyof CommonOptions>(
@@ -17,7 +22,7 @@ function getOption<K extends keyof CommonOptions>(
 export const createPersistedStatePlugin = (
   options?: PluginOptions,
 ): PiniaPlugin => {
-  const defaultStorage: Pick<Storage, 'getItem' | 'setItem' | 'removeItem'> =
+  const defaultStorage: IStorage =
     typeof window === 'object'
       ? window.localStorage
       : {
@@ -29,34 +34,47 @@ export const createPersistedStatePlugin = (
     storage: Required<CommonOptions>['storage'],
   ) => {
     const uniqueKey = '@@'
-    storage.setItem(uniqueKey, '1')
-    storage.removeItem(uniqueKey)
+    const result = storage.setItem(uniqueKey, '1')
+    if (result instanceof Promise) {
+      result.then(() => storage.removeItem(uniqueKey))
+    } else {
+      storage.removeItem(uniqueKey)
+    }
   }
 
   const pluginOptions = options ?? ({} as PluginOptions)
-  if (__DEV__ || __TEST__) {
-    ;(pluginOptions.assertStorage ?? defaultAssertStorage)(
-      pluginOptions.storage ?? defaultStorage,
-    )
-  }
 
   function plugin(context: PiniaPluginContext) {
+    // initialize custom properties
+    let resolveIsReady = () => {}
+    const isReadyPromise = new Promise<void>((resolve) => {
+      resolveIsReady = resolve
+    })
+
+    let pendingCount = 0
+    context.store.$persistedState = {
+      isReady: () => isReadyPromise,
+      pending: false,
+    }
+
     // normalize
     const options = context.options?.persistedState ?? {}
+    if (getOption(pluginOptions, options, 'persist', true) === false) return
+
     const key = options.key ?? context.store.$id
     const overwrite = getOption(pluginOptions, options, 'overwrite', false)
     const storage = getOption(pluginOptions, options, 'storage', defaultStorage)
     const filter = getOption(pluginOptions, options, 'filter', () => true)
-    const serialization = getOption(
+    const serialize = getOption(
       pluginOptions,
       options,
-      'serialization',
+      'serialize',
       JSON.stringify,
     )
-    const deserialization = getOption(
+    const deserialize = getOption(
       pluginOptions,
       options,
-      'deserialization',
+      'deserialize',
       JSON.parse,
     )
 
@@ -66,43 +84,75 @@ export const createPersistedStatePlugin = (
 
     // hydrate
     try {
-      const value = storage.getItem(key)
+      const patch = (value: any) => {
+        if (value != null) {
+          const state = deserialize(value)
 
-      if (value !== null) {
-        const state = deserialization(value)
-
-        if (overwrite) {
-          context.store.$state = state
-        } else {
-          context.store.$patch(state)
+          if (overwrite) {
+            context.store.$state = state
+          } else {
+            context.store.$patch(state)
+          }
         }
+      }
+
+      let value = storage.getItem(key)
+      if (value instanceof Promise) {
+        value.then((value) => {
+          patch(value)
+          resolveIsReady()
+        })
+      } else {
+        patch(value)
+        resolveIsReady()
       }
     } catch (error) {
       if (__DEV__ || __TEST__) console.warn(error)
+      resolveIsReady()
     }
 
     // persist
-    context.store.$subscribe((mutation, state) => {
-      if (filter(mutation, state) === false) return
+    context.store.$subscribe(
+      (
+        mutation: SubscriptionCallbackMutation<StateTree>,
+        state: {
+          [x: string]: any
+          [x: number]: any
+          [x: symbol]: any
+        },
+      ) => {
+        if (filter(mutation, state) === false) return
 
-      if (Array.isArray(options.includePaths)) {
-        state = options.includePaths.reduce((partialState, path) => {
-          return shvl.set(
-            partialState,
-            path,
-            shvl.get(state as Record<string, unknown>, path),
-          )
-        }, {})
-      } else if (Array.isArray(options.excludePaths)) {
-        state = deserialization(serialization(state))
-        options.excludePaths.forEach((path) => {
-          return shvl.set(state, path, undefined)
-        }, {})
-      }
+        if (Array.isArray(options.includePaths)) {
+          state = options.includePaths.reduce((partialState, path) => {
+            return shvl.set(
+              partialState,
+              path,
+              shvl.get(state as Record<string, unknown>, path),
+            )
+          }, {})
+        }
+        if (Array.isArray(options.excludePaths)) {
+          state = deserialize(serialize(state))
+          options.excludePaths.forEach((path) => {
+            return shvl.set(state, path, undefined)
+          }, {})
+        }
 
-      const value = serialization(state)
-      storage.setItem(key, value)
-    })
+        const value = serialize(state)
+        const result = storage.setItem(key, value)
+        if (result instanceof Promise) {
+          ++pendingCount
+          context.store.$persistedState.pending = pendingCount !== 0
+          result
+            .catch(() => {})
+            .finally(() => {
+              --pendingCount
+              context.store.$persistedState.pending = pendingCount !== 0
+            })
+        }
+      },
+    )
   }
 
   return plugin