+++ tests/index.test.ts
@@ -1,4 +1,4 @@
-import { ref, nextTick } from 'vue-demi'
+import { ref, nextTick } from 'vue'
 import { createPinia, defineStore, Pinia } from 'pinia'
 import { mount } from '@vue/test-utils'
 import { createPersistedStatePlugin } from '../src'
@@ -104,7 +104,7 @@ describe('hydrate', () => {
     spyWarn.mockRestore()
   })
 
-  it('applies deserialization when hydrating', () => {
+  it('applies deserialize when hydrating', () => {
     const spyGetItem = jest
       .spyOn(storage, 'getItem')
       .mockImplementation(() => JSON.stringify({ count: 1 }))
@@ -120,7 +120,7 @@ describe('hydrate', () => {
           count: ref(0),
         }
       },
-      { persistedState: { deserialization: spyDeserialization } },
+      { persistedState: { deserialize: spyDeserialization } },
     )()
 
     expect(spyGetItem).toBeCalled()
@@ -278,6 +278,48 @@ describe('persist', () => {
     spyWarn.mockRestore()
   })
 
+  it('persist partial state when given both of includePaths and excludePaths', async () => {
+    const spyWarn = jest.spyOn(console, 'warn').mockImplementation()
+
+    const store = defineStore(
+      'store',
+      () => {
+        return {
+          foo: ref(0),
+          bar: ref(0),
+          nested: ref({
+            baz: 0,
+            qux: 0,
+          }),
+        }
+      },
+      {
+        persistedState: {
+          includePaths: ['foo', 'nested'],
+          excludePaths: ['bar', 'nested.qux'],
+        },
+      },
+    )()
+
+    store.$state = {
+      foo: 1,
+      bar: 1,
+      nested: {
+        baz: 1,
+        qux: 1,
+      },
+    }
+    await nextTick()
+
+    expect(setItem).lastCalledWith(
+      'store',
+      JSON.stringify({ foo: 1, nested: { baz: 1 } }),
+    )
+    expect(spyWarn).not.toBeCalled()
+
+    spyWarn.mockRestore()
+  })
+
   it('not persist state when given includePaths is empty', async () => {
     const spyWarn = jest.spyOn(console, 'warn').mockImplementation()
 
@@ -386,7 +428,7 @@ describe('persist', () => {
     spyWarn.mockRestore()
   })
 
-  it('applies serialization when persisting', async () => {
+  it('applies serialize when persisting', async () => {
     const spySerialization = jest.fn().mockImplementation(() =>
       JSON.stringify({
         count: 2,
@@ -401,7 +443,7 @@ describe('persist', () => {
           count: ref(0),
         }
       },
-      { persistedState: { serialization: spySerialization } },
+      { persistedState: { serialize: spySerialization } },
     )()
     setItem.mockClear()
 
@@ -472,5 +514,99 @@ describe('persist', () => {
     await nextTick()
 
     expect(setItem).lastCalledWith(customKey, JSON.stringify({ foo: 1 }))
+  })
+
+  it('should not persist when set persist to false', async () => {
+    const store = defineStore(
+      'store',
+      () => {
+        return {
+          foo: ref(0),
+        }
+      },
+      { persistedState: { persist: false } },
+    )()
+    setItem.mockClear()
+
+    store.$state = {
+      foo: 1,
+    }
+    await nextTick()
+
+    expect(setItem).not.toBeCalled()
+  })
+
+  it('should not rehydrate when set persist to false', async () => {
+    const store = defineStore(
+      'store',
+      () => {
+        return {
+          foo: ref(0),
+        }
+      },
+      { persistedState: { persist: false } },
+    )()
+
+    expect(getItem).not.toBeCalled()
+  })
+})
+
+describe('asynchronous storage', () => {
+  beforeEach(() => {
+    const savedState: Record<string, string> = {
+      'counter-store': JSON.stringify({
+        count: 42,
+      }),
+    }
+    getItem.mockImplementation(async (key) => {
+      return savedState[key]
+    })
+    setItem.mockImplementation(async (key, value) => {
+      savedState[key] = value
+    })
+    removeItem.mockImplementation(async (key) => {
+      delete savedState[key]
+    })
+
+    pinia = createPinia().use(
+      createPersistedStatePlugin({
+        storage,
+      }),
+    )
+    mount({ template: 'none' }, { global: { plugins: [pinia] } })
+  })
+
+  it('should add $persistedState properties to store', async () => {
+    const counterStore = useCounterStore()
+
+    expect(counterStore.$persistedState).toBeInstanceOf(Object)
+    expect(counterStore.$persistedState.isReady()).toBeInstanceOf(Promise)
+    expect(typeof counterStore.$persistedState.pending).toBe('boolean')
+  })
+
+  test('$persistedState.isReady should be resolved after hydrating state', async () => {
+    const counterStore = useCounterStore()
+    expect(counterStore.count).toBe(0)
+
+    await counterStore.$persistedState.isReady()
+
+    expect(counterStore.count).toBe(42)
+  })
+
+  test('$persistedState.pending should be true while persisting state', async () => {
+    const counterStore = useCounterStore()
+
+    await counterStore.$persistedState.isReady()
+    setItem.mockClear()
+
+    counterStore.count++
+    expect(counterStore.$persistedState.pending).toBe(true)
+
+    await new Promise((resolve) => setTimeout(resolve))
+    expect(setItem).toBeCalledWith(
+      'counter-store',
+      JSON.stringify({ count: 43 }),
+    )
+    expect(counterStore.$persistedState.pending).toBe(false)
   })
 })