+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.2.0](https://github.com/iendeavor/pinia-plugin-persistedstate-2/compare/v0.1.4...v0.2.0) (2021-12-14)
+
+
+### ⚠ BREAKING CHANGES
+
+* serialization is renamed as serialize, deserialization is renamed as deserialize
+
+### Features
+
+* allow to disable plugin by default and enable for specific stores ([347d51c](https://github.com/iendeavor/pinia-plugin-persistedstate-2/commit/347d51c6c37d50ad2afd9a0e0f9c24dbcf58ef32))
+* allow to persist blob or other types supported by storage ([f23ace7](https://github.com/iendeavor/pinia-plugin-persistedstate-2/commit/f23ace787f48d90e83f8c845aebe0e4022c25ce8))
+* use verb ([106ba31](https://github.com/iendeavor/pinia-plugin-persistedstate-2/commit/106ba31e01ece9a80484d6f50d6049eedf18436c))
+
+
+### Bug Fixes
+
+* exclude path may be in include path ([4e6c04f](https://github.com/iendeavor/pinia-plugin-persistedstate-2/commit/4e6c04f713a5ce08c11d6da5c2560f706d5a716b))
+
+### [0.1.4](https://github.com/iendeavor/pinia-plugin-persistedstate-2/compare/v0.1.3...v0.1.4) (2021-12-11)
+
+
+### Features
+
+* asynchronous storage ([a5f6097](https://github.com/iendeavor/pinia-plugin-persistedstate-2/commit/a5f6097bb067a4f74ecded52265bbe38f5a74a4d)), closes [#3](https://github.com/iendeavor/pinia-plugin-persistedstate-2/issues/3)
+
+### [0.1.3](https://github.com/iendeavor/pinia-plugin-persistedstate-2/compare/v0.1.2...v0.1.3) (2021-12-10)
+
+
+### Features
+
+* allow to disable persisting stores ([34a0a06](https://github.com/iendeavor/pinia-plugin-persistedstate-2/commit/34a0a06dfc41c3eeced675ba8f495d07d422c824))
+
 ### [0.1.2](https://github.com/iendeavor/pinia-plugin-persistedstate-2/compare/v0.1.1...v0.1.2) (2021-12-05)